pipecrowd/requirements.txt
Django==3.0
djangorestframework==3.10.3
google-cloud-storage==1.24.1
psycopg2-binary==2.8.4
psycopg2==2.8.4
django-cors-headers==3.2.1